Rename ftml classes.
