--- web/app/Services/Wikitext/FtmlBackend.php
@@ -1,16 +1,0 @@
-<?php
-declare(strict_types = 1);
-
-namespace Wikidot\Wikitext\FTML;
-
-use \FFI;
-
-/**
- * Class FtmlBackend, implements a compatible interface for working with FTML.
- * @package Wikidot\Wikitext\FTML
- */
-class FtmlBackend
-{
-    // TODO -- implement WikiTransformation-like interface
-    // see https://scuttle.atlassian.net/browse/WJ-463
-}
