+++ web/app/Services/Wikitext/HtmlOutput.php
@@ -1,15 +1,15 @@
 <?php
 declare(strict_types = 1);
 
-namespace Wikidot\Wikitext\FTML;
+namespace Wikijump\Services\Wikitext;
 
 use \FFI;
 
 /**
- * Class FtmlHtmlOutput, representing a returned 'struct ftml_html_output' object.
- * @package Wikidot\Wikitext\FTML
+ * Class HtmlOutput, representing a returned 'struct ftml_html_output' object.
+ * @package Wikijump\Services\Wikitext
  */
-class FtmlHtmlOutput
+class HtmlOutput
 {
     public string $html;
     public string $style;
@@ -19,8 +19,8 @@ class FtmlHtmlOutput
     public function __construct(FFI\CData $c_data) {
         $this->html = FFI::string($c_data->html);
         $this->style = FFI::string($c_data->style);
-        $this->meta = FtmlHtmlMeta::fromArray($c_data->meta_list, $c_data->meta_len);
-        $this->warnings = FtmlWarning::fromArray($c_data->warning_list, $c_data->warning_len);
+        $this->meta = HtmlMeta::fromArray($c_data->meta_list, $c_data->meta_len);
+        $this->warnings = ParseWarning::fromArray($c_data->warning_list, $c_data->warning_len);
 
         // Free original C data
         FtmlFfi::freeHtmlOutput($c_data);
