+++ web/app/Services/Wikitext/WikitextInterface.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types = 1);
+
+namespace Wikijump\Services\Wikitext;
+
+use \FFI;
+
+/**
+ * Class WikitextInterface, implements a compatible interface for working with FTML.
+ * @package Wikijump\Services\Wikitext
+ */
+class WikitextInterface
+{
+    // TODO -- implement WikiTransformation-like interface
+    // see https://scuttle.atlassian.net/browse/WJ-463
+}
